clean compile warning (#2380)

+++ core/src/db/merge/MergeAdaptiveStrategy.cpp
@@ -32,7 +32,7 @@ MergeAdaptiveStrategy::RegroupFiles(meta::FilesHolder& files_holder, MergeFilesG
     meta::SegmentsSchema& files = files_holder.HoldFiles();
     for (meta::SegmentsSchema::reverse_iterator iter = files.rbegin(); iter != files.rend(); ++iter) {
         meta::SegmentSchema& file = *iter;
-        if (file.index_file_size_ > 0 && file.file_size_ > file.index_file_size_) {
+        if (file.index_file_size_ > 0 && (int64_t)file.file_size_ > file.index_file_size_) {
             // file that no need to merge
             continue;
         }
@@ -60,7 +60,7 @@ MergeAdaptiveStrategy::RegroupFiles(meta::FilesHolder& files_holder, MergeFilesG
         int64_t sum_size = 0;
         for (auto iter = sort_files.begin(); iter != sort_files.end();) {
             meta::SegmentSchema& file = *iter;
-            if (sum_size + file.file_size_ <= index_file_size) {
+            if (sum_size + (int64_t)(file.file_size_) <= index_file_size) {
                 temp_group.push_back(file);
                 sum_size += file.file_size_;
                 iter = sort_files.erase(iter);
